refactor: update header step logic and clean up files
