--- src/NextGen/DonationForm/resources/app/form/MultiStepForm/components/StepHeader.tsx
@@ -1,7 +1,0 @@
-import useCurrentStep from '@givewp/forms/app/form/MultiStepForm/hooks/useCurrentStep';
-
-export default function StepHeader() {
-    const step = useCurrentStep();
-
-    return step.id > 0 && <p>{step.title}</p>;
-}
