+++ src/NextGen/DonationForm/resources/app/form/MultiStepForm/components/SubmitButton.tsx
@@ -1,9 +1,12 @@
 import {__} from '@wordpress/i18n';
 
+/**
+ * @unreleased
+ */
 export default function SubmitButton({
-    isSubmitting,
-    submittingText = __('Submitting…', 'give'),
-    buttonText = __('Donate Now', 'give'),
+                                         isSubmitting,
+                                         submittingText = __('Submitting…', 'give'),
+                                         buttonText = __('Donate Now', 'give'),
 }) {
     return (
         <button type="submit" disabled={isSubmitting} aria-busy={isSubmitting}>
