+++ src/NextGen/DonationForm/resources/app/form/MultiStepForm/types/index.ts
@@ -1,5 +1,8 @@
 import {ReactElement} from 'react';
 
+/**
+ * @unreleased
+ */
 export type FormInputs = {
     FORM_ERROR: string;
     amount: number;
@@ -9,6 +12,9 @@ export type FormInputs = {
     gatewayId: string;
 };
 
+/**
+ * @unreleased
+ */
 export type StepObject = {
     id: number;
     title: string;
