+++ src/NextGen/DonationForm/resources/app/form/MultiStepForm/utilities/getCurrentStepObject.ts
@@ -4,5 +4,5 @@ import {StepObject} from "@givewp/forms/app/form/MultiStepForm/types";
  * @unreleased
  */
 export default function getCurrentStepObject(steps: StepObject[], currentStep: number) {
-    return steps.find(({id}) => id === currentStep);
+    return steps?.find(({id}) => id === currentStep);
 }
