+++ src/NextGen/DonationForm/resources/app/store/index.tsx
@@ -1,5 +1,5 @@
 import {createContext, ReactNode, useContext, useReducer} from 'react';
-import type {Gateway, Section} from '@givewp/forms/types';
+import type {Gateway} from '@givewp/forms/types';
 import reducer from '@givewp/forms/app/store/reducer';
 import {ObjectSchema} from 'joi';
 
@@ -12,7 +12,6 @@ StoreContextDispatch.displayName = 'DonationFormStateDispatch';
 type PropTypes = {
     initialState: {
         gateways: Gateway[];
-        sections: Section[];
         defaultValues: object;
         validationSchema: ObjectSchema;
     };
